infra/requirements.txt
attrs==20.3.0
aws-cdk.assets==1.90.1
aws-cdk.aws-applicationautoscaling==1.90.1
aws-cdk.aws-autoscaling-common==1.90.1
aws-cdk.aws-cloudwatch==1.90.1
aws-cdk.aws-codeguruprofiler==1.90.1
aws-cdk.aws-ec2==1.90.1
aws-cdk.aws-ecr==1.90.1
aws-cdk.aws-ecr-assets==1.90.1
aws-cdk.aws-efs==1.90.1
aws-cdk.aws-events==1.90.1
aws-cdk.aws-iam==1.90.1
aws-cdk.aws-kms==1.90.1
aws-cdk.aws-lambda==1.90.1
aws-cdk.aws-logs==1.90.1
aws-cdk.aws-s3==1.90.1
aws-cdk.aws-s3-assets==1.90.1
aws-cdk.aws-sqs==1.90.1
aws-cdk.aws-ssm==1.90.1
aws-cdk.aws-stepfunctions==1.90.1
aws-cdk.cloud-assembly-schema==1.90.1
aws-cdk.core==1.90.1
aws-cdk.cx-api==1.90.1
aws-cdk.region-info==1.90.1
cattrs==1.1.2
configparser==5.0.1
constructs==3.3.32
jsii==1.21.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
